serviceplatform/requirements.txt
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
Django==1.11.4
django-filter==1.1.0
django-guardian==1.4.9
django-mssql==1.8
django-pyodbc==1.1.3
django-pytds==1.5
django-simple-rest==1.4.1
django-sqlserver==1.11
DjangoCaptcha==0.3.5
djangorestframework==3.8.2
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
httpie==0.9.9
httplib2==0.10.3
idna==2.5
itchat==1.3.10
JPype1==0.6.3
Keras==2.1.2
Markdown==2.6.11
mimeparse==0.1.3
netaddr==0.7.18
numpy==1.14.0
olefile==0.44
Pillow==4.3.0
protobuf==3.5.1
Pygments==2.2.0
pymongo==2.9
PyMySQL==0.7.11
pyodbc==4.0.23
pypiwin32==223
pypng==0.0.18
PyQRCode==1.2.1
python-tds==1.9.1
pytz==2017.2
pywin32==223
PyYAML==3.12
requests==2.17.3
scipy==1.0.0
selenium==3.11.0
sh==1.12.14
simplejson==3.14.0
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
tornado==4.2.1
urllib3==1.21.1
Werkzeug==0.14.1
